# [lcm_gen] Optimize encoding and decoding of arrays