As is turns out, the compiler doesn't inline recursive calls to encode or decode, even though the call graph and constexpr branches are known ahead of time. Instead, we'll directly check for the case of copying contiguous memory with a special case.

```
                     lcm-gen  lcm_gen  lcm_gen
                    upstream   before    after
ImageArrayEncode     77.9 us   975 us  78.3 us
ImageArrayDecode     80.0 us   979 us  81.3 us
PandaCommandEncode   29.3 ns  21.0 ns  20.9 ns
PandaCommandDecode   3.12 ns  3.08 ns  3.10 ns
PandaStatusEncode    69.2 ns  43.0 ns  34.9 ns
PandaStatusDecode    3.11 ns  3.08 ns  3.10 ns
ViewerDrawEncode     4.97 us  3.42 us  2.53 us
ViewerDrawDecode     3.11 ns  3.22 ns  3.09 ns
```

Note that by default all of our messages still use the lcm-gen (legacy upstream) message generator. This commit is just fixing performance prior to the switch-over.

---

Towards #20761.

The bottom line of the benchmark above is that this PR ("lcm_gen after") is on par with the performance of upstream's `lcm-gen` program.  (It's of course way better than our current on-master implementation "lcm_gen before".)

I have a few more micro-optimizations on deck after this, but this one alone is sufficient to unblock us from switching our default over to the new implementation.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/RobotLocomotion/drake/22525)
<!-- Reviewable:end -->